+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+#### 0.1.3
+- Fix [#7]([https://github.com/jolleekin/modern_charts/issues/7)
+- Fix [#8]([https://github.com/jolleekin/modern_charts/issues/8)
+- Add `yAxis`:`minInterval` setting to BarChart, LineChart, and RadarChart
+- Perform code cleanup
+
 #### 0.1.2
 - Fix [#5]([https://github.com/jolleekin/modern_charts/issues/5)
 - Fix [#6]([https://github.com/jolleekin/modern_charts/issues/6)